Add files via upload
An assignment I completed for my C++ class at CCAC. It takes any double as input and applies the Babylonian square root algorithm until a number within 0.001 of the correct answer is found. This can be done any number of times the user desires. 

Limitations of this program: very small decimal numbers less than 1 do not work well.

The properties of this file will say that it was made recently, but the date typed at the top of the file is when the code itself was written.