Add files via upload
This program was an assignment for my C++ course. It takes a long long integer input and "rolls" a pair of six-sided "dice" that many times. After the rolling is complete, the number of times a given possible result occurred is shown and compared to the calculated odds, or expected # of times, of it occurring within a table. The % error between the expected and actual number of rolls is then calculated and tabulated as well. The user can repeat this process as many times as desired.